[fix/managingMember] 멤버 추방 api transactional 메서드 단에 적용

+++ src/main/java/GDG/whatssue/domain/member/service/ClubMemberManagingService.java
@@ -20,10 +20,15 @@ public class ClubMemberManagingService {
 
     private final ClubMemberRepository clubMemberRepository;
 
-    public void deleteClubMember(Long memberId) {
+    @Transactional
+    public void deleteClubMember(Long clubId, Long memberId) {
 
         ClubMember clubMember = clubMemberRepository.findById(memberId)
             .orElseThrow(() -> new CommonException(ClubMemberErrorCode.EX2100));
+
+        if (clubMember.getClub().getId() != clubId){
+            throw new CommonException(ClubMemberErrorCode.EX2203);
+        }
 
         clubMemberRepository.delete(clubMember);
     }
